backend/evaluation/pyproject.toml
[project]
name = "fraud-detection-evaluation"
version = "1.0.0"
description = "RAG evaluation service with Pinecone, TruLens, and RAGAS"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pinecone>=5.0.0",
    "trulens>=1.0.0",
    "trulens-providers-litellm>=1.0.0",
    "ragas>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-pinecone>=0.2.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "datasets>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "httpx>=0.27.0"]
vector = ["qdrant-client>=1.12.0", "sentence-transformers>=3.0.0", "chromadb>=0.5.0", "weaviate-client>=4.0.0"]
memory = ["mem0ai>=0.1.0"]
deepeval = ["deepeval>=1.0.0"]
braintrust = ["braintrust>=0.0.100"]
letta = ["letta>=0.6.0"]
phoenix = ["arize-phoenix>=8.0.0", "opentelemetry-api>=1.0.0", "opentelemetry-sdk>=1.0.0"]

[tool.setuptools]
py-modules = ["config", "main"]
